Combine multiple network calls into one [BE-1576]

See merge request beamgroup/beam!1058

+++ Beam/Classes/Components/DocumentsList/DocumentDetail.swift
@@ -238,7 +238,7 @@ struct DocumentDetail: View {
 
     static private let dateFormat: DateFormatter = {
         let formatter = DateFormatter()
-        formatter.dateStyle = .short
+        formatter.dateStyle = .long
         formatter.timeStyle = .long
         return formatter
     }()
